removed abstract definition; moved to resources namespace
